Fix subscription billing duration scaling and payment amount type
- Add billing_duration to UserResponse schema for frontend access
- Set billing_duration on user when processing payments (Razorpay & Dodo)
- Fix dashboard to show scaled limits (quarterly=3x, yearly=12x)
- Update usage label to show "Quarter/Half-Year/Year" based on plan
- Change Payment.amount from Integer to Float for decimal precision
- Add database migrations for billing_duration and amount type change

+++ backend/app/models/payment.py
@@ -5,7 +5,7 @@ This module defines the Payment database model for tracking Razorpay
 payment transactions and subscription purchases.
 """
 
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Enum
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Enum, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..database import Base
@@ -59,7 +59,7 @@ class Payment(Base):
     dodo_payment_id = Column(String(255), nullable=True)
 
     # Payment details
-    amount = Column(Integer, nullable=False)  # Amount in paise
+    amount = Column(Float, nullable=False)  # Amount in smallest currency unit (paise/cents)
     currency = Column(String, default="INR", nullable=False)
     status = Column(
         Enum(PaymentStatus),
